requirements.txt
fastapi[standard]
fastapi-users[sqlalchemy]
uvicorn[standard]
psycopg[binary,pool]
asyncpg
alembic
python-dotenv